Add an example of how a library writer can handle user-defined callback.
On the library side, the engine use TLS to store the event handler which will be called by a TLS
lookup on the glfw callback. It stores user callbacks on the event handler as @fn.

On the user side, user-defined callbacks are registered as @fn, every user data shared by several
callbacks have to be @-boxed.